Don't desguar Map to MapAccess for unknown funcs
Maps to functions we don't know about should be
passed as references by default.


stack-info: PR: https://github.com/bpftrace/bpftrace/pull/4452, branch: user/jordalgo/addr/2
